+++ App.tsx
@@ -1,52 +1,28 @@
 import React from 'react';
-import { SafeAreaView, StyleSheet, View, Image } from 'react-native';
-import Button from './src/components/Button';
-import Input from './src/components/Input';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import Login from './src/pages/Login';
+import Home from './src/pages/Home';
 
+const Stack = createNativeStackNavigator();
 
 function App(): React.JSX.Element {
-
-  const handleAvancarPress = () => {
-    console.log('Botão Avançar pressionado!');
-  };
-
   return (
-    <SafeAreaView style={styles.container}>
-      <Image
-        source={require('./src/assets/LOGO.png')}
-        style={styles.logo}
-        resizeMode="contain"
-      />
-
-      <Input label="Login:" placeholder="Digite seu login" secureTextEntry={false}/>
-      <Input label="Senha:" placeholder="Digite sua senha" secureTextEntry={true}/>
-
-      <View style={styles.buttonContainer}>
-        <Button color="yellow" onPress={handleAvancarPress}>
-          Entrar
-        </Button>
-      </View>
-    </SafeAreaView>
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Login">
+        <Stack.Screen 
+          name="Login" 
+          component={Login} 
+          options={{ headerShown: false }}
+        />
+        <Stack.Screen 
+          name="Home" 
+          component={Home}
+          options={{ headerShown: true }}
+        />
+      </Stack.Navigator>
+    </NavigationContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingHorizontal: 20,
-  },
-  logo: {
-    width: 200,
-    height: 200,
-    marginBottom: 20,
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    marginTop: 20,
-  },
-});
 
 export default App;
