+++ package.json
@@ -11,12 +11,16 @@
   },
   "dependencies": {
     "@react-native-community/geolocation": "^3.4.0",
+    "@react-navigation/native": "^6.1.18",
+    "@react-navigation/native-stack": "^6.11.0",
     "@types/react-native": "^0.73.0",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
     "react": "18.3.1",
     "react-native": "0.75.3",
     "react-native-maps": "^1.18.0",
+    "react-native-safe-area-context": "^4.11.0",
+    "react-native-screens": "^3.34.0",
     "react-native-vector-icons": "^10.2.0"
   },
   "devDependencies": {
